refactor(collab): addressed pr comments
