+++ vitest.config.ts
@@ -3,7 +3,6 @@ import { defineConfig } from "vitest/config";
 // https://vitejs.dev/config/
 export default defineConfig({
     test: {
-        testTimeout: 40000,
         alias: {
             "react/jsx-dev-runtime": "preact/jsx-runtime",
         },
